Finished exercises 1 and 2

+++ List-1/temperature-converter/solution.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
-int converter(float celsius){
+float converter(float celsius){
 	float farenheit = (celsius*9/5)+32;	
 	return farenheit;
 }
 
 void main(){
-        printf("Temperature is: %f \n", converter(32));
+    printf("Temperature is: %0.2f \n", converter(32));
 }
